update: restructure like JS
